style: restructured finance template and it's components's respective sass files
