+++ src/Templates/Finance/Pages/Finance.tsx
@@ -1,7 +1,7 @@
 import '../../../assets/styles/templatesStyles/Finance/Finance.scss'
 import TemplateData from '../../../staticDB/FinanceData'
-import searchlogo from '../assets/images/Vector (5).png'
-import puzzleImg from '../assets/images/puzzle.png'
+import searchlogo from '../../../assets/images/Vector (5).png'
+import puzzleImg from '../../../assets/images/puzzle.png'
 import React, { useState } from 'react'
 import { MessageForm } from '../Components/MessageForm'
 import { Footer } from '../Components/Footer'
