+++ src/assets/styles/templatesStyles/Finance/MessageForm.scss
@@ -1,17 +1,20 @@
 .finance-form-div{
+   
     display: flex;
     width: 28rem;
-    height: 74%;
+    height: 25rem;
     .form{
         display: flex;
         flex-direction: column;
+        align-items: flex-start;
+        gap: 2rem;
         width: 100%;
         height: 100%;
         input[type="text"]{
-            margin-bottom: 2rem;
-            width: 95;
+            width: 95%;
+            margin: 0;
             position: relative;
-            height: 2.5rem;
+            height: 2.8rem;
             border-radius: 10px;
             border: none;
             background: #FFFFFF;
@@ -25,14 +28,19 @@
             outline: none;
         }
         label{
+            position: relative;
+            top: 1.7rem;
             font-family: 'poppins';
             font-weight: 400;
             font-size: 16px;
+            margin: 0;
             color:  #484848;
         }
         textarea{
             border-radius: 10px;
             padding: 10px;
+            margin: 0;
+            width: 90%;
             font-weight: 400;
             color: #CACACA;
             border: none;
@@ -43,14 +51,15 @@
         }
         .submit-button{
             width: 6.5rem;
+            position: relative;
+            bottom: 1rem;
             height: 2.2rem;
             border: none;
             cursor: pointer;
             background-color: #222222;
             color: #FFFFFF;
             border-radius: 10px;
-            margin-top: 1.5rem;
-        
+            margin: 0;
         }
     }
 }
